move session-related code into SesisonHelper
